Print json of inputted room id to /join-room page
