+++ pom.xml
@@ -19,10 +19,20 @@
   <dependencies>
     <dependency>
       <groupId>javax.servlet</groupId>
-      <artifactId>javax.servlet-api</artifactId>
-      <version>4.0.1</version>
-      <scope>provided</scope>
+    	<artifactId>javax.servlet-api</artifactId>
+    	<version>4.0.1</version>
+    	<scope>provided</scope>
     </dependency>
+    <dependency>
+    	<groupId>com.google.code.gson</groupId>
+      <artifactId>gson</artifactId>
+    	<version>2.8.6</version>
+    </dependency>
+    <dependency>
+    	<groupId>com.google.appengine</groupId>
+    	<artifactId>appengine-api-1.0-sdk</artifactId>
+    	<version>1.9.59</version>
+   	</dependency>
   </dependencies>
 
   <build>
