+++ src/main/webapp/join-room.html
@@ -13,7 +13,7 @@
   </head>
   <body>
     <div id="join-room-box">
-      <form action="/join-room" method="GET">
+      <form action="/join-room" method="POST">
         <h2>Input your private room link Below!</h2>
         <input type="text" name="user-party-link">
         <input type="submit"/>
